alex/arrows-label-placement: cleanup
