+++ packages/editor/src/index.ts
@@ -309,6 +309,7 @@ export {
 	PI,
 	PI2,
 	SIN,
+	angleDistance,
 	approximately,
 	areAnglesCompatible,
 	average,
@@ -318,8 +319,6 @@ export {
 	clockwiseAngleDist,
 	counterClockwiseAngleDist,
 	degreesToRadians,
-	getArcMeasure,
-	getPointInArcT,
 	getPointOnCircle,
 	getPolygonVertices,
 	isSafeFloat,
