+++ packages/editor/src/lib/primitives/utils.ts
@@ -358,47 +358,17 @@ export const isSafeFloat = (n: number) => {
 }
 
 /**
- * Returns the t value of the point on the arc.
- *
- * @param mAB - The measure of the arc from A to B, negative if counter-clockwise
- * @param A - The angle from center to arc's start point (A) on the circle
- * @param B - The angle from center to arc's end point (B) on the circle
- * @param P - The angle on the circle (P) to find the t value for
- *
- * @returns The t value of the point on the arc, with 0 being the start and 1 being the end
- *
+ * Get the angle of a point on an arc.
+ * @param fromAngle - The angle from center to arc's start point (A) on the circle
+ * @param toAngle - The angle from center to arc's end point (B) on the circle
+ * @param direction - The direction of the arc (1 = counter-clockwise, -1 = clockwise)
+ * @returns The distance in radians between the two angles according to the direction
  * @public
  */
-export function getPointInArcT(mAB: number, A: number, B: number, P: number): number {
-	let mAP: number
-	if (Math.abs(mAB) > PI) {
-		mAP = shortAngleDist(A, P)
-		const mPB = shortAngleDist(P, B)
-		if (Math.abs(mAP) < Math.abs(mPB)) {
-			return mAP / mAB
-		} else {
-			return (mAB - mPB) / mAB
-		}
-	} else {
-		mAP = shortAngleDist(A, P)
-		return mAP / mAB
-	}
-}
-
-/**
- * Get the measure of an arc.
- *
- * @param A - The angle from center to arc's start point (A) on the circle
- * @param B - The angle from center to arc's end point (B) on the circle
- * @param sweepFlag - 1 if the arc is clockwise, 0 if counter-clockwise
- * @param largeArcFlag - 1 if the arc is greater than 180 degrees, 0 if less than 180 degrees
- *
- * @returns The measure of the arc, negative if counter-clockwise
- *
- * @public
- */
-export function getArcMeasure(A: number, B: number, sweepFlag: boolean, largeArcFlag: boolean) {
-	const m = ((2 * ((B - A) % PI2)) % PI2) - ((B - A) % PI2)
-	if (!largeArcFlag) return m
-	return (PI2 - Math.abs(m)) * (sweepFlag ? 1 : -1)
+export function angleDistance(fromAngle: number, toAngle: number, direction: number) {
+	const dist =
+		direction < 0
+			? clockwiseAngleDist(fromAngle, toAngle)
+			: counterClockwiseAngleDist(fromAngle, toAngle)
+	return dist
 }
